staging/pkg/mod/golang.org/x/net@v0.0.0-20201110031124-69a78807bb2b/html/testdata/go/raw_tags_to_be_ignored.dat

#data
<!doctype html><table><select><iframe>a<caption>b
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       "a"
|     <table>
|       <caption>
|         "b"

#data
<!doctype html><table><select><noembed>a<caption>b
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       "a"
|     <table>
|       <caption>
|         "b"

#data
<!doctype html><table><select><noframes>a<caption>b
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       "a"
|     <table>
|       <caption>
|         "b"

#data
<!doctype html><table><select><noscript>a<caption>b
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       "a"
|     <table>
|       <caption>
|         "b"

#data
<!doctype html><table><select><style>a<caption>b
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       "a"
|     <table>
|       <caption>
|         "b"

#data
<!doctype html><table><select><title>a<caption>b
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       "a"
|     <table>
|       <caption>
|         "b"

#data
<!doctype html><table><select><xmp>a<caption>b
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       "a"
|     <table>
|       <caption>
|         "b"
